Move to double duration

+++ submodules/AccountContext/Sources/UniversalVideoNode.swift
@@ -37,7 +37,7 @@ public protocol UniversalVideoContentNode: AnyObject {
 public protocol UniversalVideoContent {
     var id: AnyHashable { get }
     var dimensions: CGSize { get }
-    var duration: Int32 { get }
+    var duration: Double { get }
     
     func makeContentNode(postbox: Postbox, audioSession: ManagedAudioSession) -> UniversalVideoContentNode & ASDisplayNode
