+++ submodules/TelegramUI/Sources/GridMessageItem.swift
@@ -272,7 +272,7 @@ final class GridMessageItemNode: GridItemNode {
                         })
                         
                         if let duration = file.duration {
-                            let durationString = stringForDuration(duration)
+                            let durationString = stringForDuration(Int32(duration))
                             
                             var badgeContent: ChatMessageInteractiveMediaBadgeContent?
                             var mediaDownloadState: ChatMessageInteractiveMediaDownloadState?
